requirements.txt
alembic==1.3.0
Babel==2.7.0
blinker==1.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-CKEditor==0.4.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
httplib2==0.14.0
humanize==0.5.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
oauth2==1.9.0.post1
passlib==1.7.1
psycopg2==2.8.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==4.0.0
pytz==2019.3
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.10
text-unidecode==1.3
tzlocal==2.0.0
Unidecode==1.1.1
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
